+++ mods/BrewLAN/units/SEA0004/SEA0004_script.lua
@@ -1,5 +1,0 @@
-local TAirUnit = import('/lua/terranunits.lua').TAirUnit
-
-SEA0004 = Class(TAirUnit) {}
-
-TypeClass = SEA0004